[add] -move socket logic from homepage to yelplist
